backend/app/agents.py
import os
from dotenv import load_dotenv

# TODO: [GEMINI INTEGRATION] 
# import google.generativeai as genai
# genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

load_dotenv()

async def mock_gemini_call(agent_name: str, context: str) -> str:
    """Mock Gemini layer returning structured agent notes."""
    # TODO: [GEMINI INTEGRATION] Replace with model.generate_content_async(context)
    return f"[{agent_name} output based on inputs: {context[:50]}...]"

async def run_equity_agent(data_json: str, e_score: float) -> str:
    prompt = "You are the Equity Agent. Assess demographic vulnerability and healthcare needs."
    return await mock_gemini_call("Equity Agent", f"Prompt: {prompt} Data: {data_json} E-Score: {e_score}")

async def run_logistics_agent(data_json: str, l_score: float) -> str:
    prompt = "You are the Logistics Agent. Assess road closures and delivery viability."
    return await mock_gemini_call("Logistics Agent", f"Prompt: {prompt} Data: {data_json} L-Score: {l_score}")

async def run_risk_agent(data_json: str, r_score: float) -> str:
    prompt = "You are the Risk Agent. Assess black market diversion likelihood."
    return await mock_gemini_call("Risk Agent", f"Prompt: {prompt} Data: {data_json} R-Score: {r_score}")

async def run_demand_agent(data_json: str, d_score: float) -> str:
    prompt = "You are the Demand Agent. Assess urgency versus induction cooktop displacement."
    return await mock_gemini_call("Demand Agent", f"Prompt: {prompt} Data: {data_json} D-Score: {d_score}")

async def run_moderator_agent(inputs_json: str, e_txt: str, l_txt: str, r_txt: str, d_txt: str) -> str:
    prompt = "You are the Moderator Agent. Resolve agent conflicts and justify the Priority Score."
    context = f"Inputs: {inputs_json} | Equity: {e_txt} | Logistics: {l_txt} | Risk: {r_txt} | Demand: {d_txt}"
    return await mock_gemini_call("Moderator Agent", f"Prompt: {prompt} Context: {context}")
