backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
google-generativeai==0.4.1
python-dotenv==1.0.1
firebase-admin==6.5.0
